gemini-python/python-server/requirements.txt
google-genai
websockets
python-dotenv
aiohttp